requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.2.1
pymysql==1.1.0
email-validator==2.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
Werkzeug==2.3.7 